refactor: review 찜목록 카드 컴포넌트 분리
